# par_bridge: use naive locking of the Iterator