@njaard 2022-12-06
@cupiver good idea

I ran the two demos that that appear to reference par_bridge, `nbody` and `life`.

# Without my patch
```
$ hyperfine "./target/release/rayon-demo  life  bench"
Benchmark 1: ./target/release/rayon-demo  life  bench
  Time (mean ± σ):     811.4 ms ±  26.4 ms    [User: 2784.5 ms, System: 81.1 ms]
  Range (min … max):   751.2 ms … 849.7 ms    10 runs

$ hyperfine "./target/release/rayon-demo  nbody bench"
Benchmark 1: ./target/release/rayon-demo  nbody bench
  Time (mean ± σ):      7.660 s ±  0.212 s    [User: 17.835 s, System: 0.044 s]
  Range (min … max):    7.416 s …  8.070 s    10 runs
```

# With my patch
```
$ hyperfine "./target/release/rayon-demo  life  bench"
Benchmark 1: ./target/release/rayon-demo  life  bench
  Time (mean ± σ):     503.0 ms ±  21.3 ms    [User: 1073.8 ms, System: 472.3 ms]
  Range (min … max):   487.0 ms … 548.7 ms    10 runs

$ hyperfine "./target/release/rayon-demo  nbody bench"
Benchmark 1: ./target/release/rayon-demo  nbody bench
  Time (mean ± σ):      7.699 s ±  0.144 s    [User: 18.002 s, System: 0.051 s]
  Range (min … max):    7.507 s …  7.922 s    10 runs
  
```

# Summary
The "nbody" task is slightly worse with my patch, but not statistically significantly. The "life" task is considerably faster with my patch.

# Machine
Performed on a 4-core CPU