@njaard 2022-12-06
# With my patch
```
test nbody::bench::nbody_par_bridge                                     ... bench:  10,193,883 ns/iter (+/- 14,547,250)
test nbody::bench::nbody_par_iter                                       ... bench:  10,313,040 ns/iter (+/- 9,173,607)
test nbody::bench::nbody_parreduce                                      ... bench:  21,930,914 ns/iter (+/- 10,491,332)
test nbody::bench::nbody_seq                                            ... bench:  37,272,289 ns/iter (+/- 34,929,418)

test life::bench::generations                                           ... bench:  73,372,701 ns/iter (+/- 6,076,736)
test life::bench::par_bridge_generations                                ... bench: 384,203,006 ns/iter (+/- 29,417,758)
test life::bench::par_iter_generations                                  ... bench:  23,879,561 ns/iter (+/- 1,021,068)
```

# Without my patch
```
test nbody::bench::nbody_par_bridge                                     ... bench:  10,263,138 ns/iter (+/- 9,227,833)
test nbody::bench::nbody_par_iter                                       ... bench:  11,594,927 ns/iter (+/- 12,503,276)
test nbody::bench::nbody_parreduce                                      ... bench:  22,593,455 ns/iter (+/- 10,537,023)
test nbody::bench::nbody_seq                                            ... bench:  36,330,253 ns/iter (+/- 35,849,257)

test life::bench::generations                                           ... bench:  71,825,838 ns/iter (+/- 3,260,082)
test life::bench::par_bridge_generations                                ... bench: 738,822,305 ns/iter (+/- 135,174,311)
test life::bench::par_iter_generations                                  ... bench:  24,250,743 ns/iter (+/- 1,219,741)
```

Looks like my patch provides a huge benefit.