@adamreichold 2022-12-07
Does the removal of `done` have any measurable effect on the benchmarks?