@cuviper 2022-12-08
Looks good, thanks! I'm going to followup with an attempt at preventing mutex recursion (#690), then we can publish.

bors r+